Fix issues with filesystem::perms in Windows-GCC (#23)
On Windows a file/dir with full access permissions seems not to have
perms::all set. This may be caused by permissions available on POSIX
but not on Windows. Instead of testing file/dirs with full access for
perms::all the test checks specifically for the permissions of
interest.

Fixing usage of gtest EXPECT_EQ for better readability of error messages.