Replace GetExpandedArgv with GetExpandedArgs
When generating the cbuild YAML, individual access to switch and value
is necessary. Also, the arguments are no longer quoted (not wanted when
producing the cbuild YAML file). The executable can be retrieved with
GetExecutable, and doesn't fit with the [switch, value] pair, so that
is no longer returned as the first argument.

See https://github.com/Open-CMSIS-Pack/devtools/pull/673#issuecomment-1415427391

Contributed by STMicroelectronics

Signed-off-by: Erik MÅLLBERG <erik.lundin@st.com>