Directories should be executable even if read-only
If directories are not marked as executable, it's not possible to list
the content. This greatly complicates navigating the "output" directory
tree when running the test suite.

From now on, reseting the read-only state on a resource will inherit
the write permission from the parent directory. If the parent directory
is not writable, then the current umask value is considered.

Contributed by STMicroelectronics

Signed-off-by: Torbjörn SVENSSON <torbjorn.svensson@foss.st.com>