Fix crash when same pack is located on 2 paths
Scenario:
csolution.yml:
packs:
  - pack: ARM::CMSIS@5.9.0
    path: ./ARM.CMSIS.5.9.0/

cproject.yml:
packs:
  - pack: ARM.CMSIS@5.9.0

In the above scenario, the same pack is defined on 2 different paths.
The one in the csolution.yml is a project local path, and for the
cproject.yml, it can be found in the $CMSIS_PACK_ROOT.

As the RteModel does not allow "duplicates", the RtePackage pointer,
for the second pdsc file to be loaded, will be released when
calling RteModel->Validate() and thus, only instances that is known
by the model should be contained in m_loadedPacks in the ProjMgrWorker
instance.

Contributed by STMicroelectronics

Signed-off-by: Torbjörn SVENSSON <torbjorn.svensson@foss.st.com>
Co-authored-by: Daniel Brondani <daniel.brondani@arm.com>