Do not validate for http:// URLs in elements and attributes that are … (#2094)
…definitely part of a different namespace

Since the XML parser does not seem to have much support at all for
namespaces, this fix is very ... raw in that it only looks for a
namespace delimiter `:` and if we find that in an element, mark it as
ok. If we find it in an attribute, do not check it.

This fixes #2093

Co-authored-by: Evgueni Driouk <edriouk@arm.com>